Resolve warnings

+++ mario-directx/Game.cpp
@@ -73,8 +73,8 @@ void CGame::Draw(float x, float y, LPDIRECT3DTEXTURE9 texture, int left, int top
 	float cam_x, cam_y;
 	CCamera::GetInstance()->GetPosition(cam_x, cam_y);
 
-	cam_x = (cam_x - (int)cam_x >= 0.5f) ? (int)cam_x + 1 : (int)cam_x;
-	cam_y = (cam_y - (int)cam_y >= 0.5f) ? (int)cam_y + 1 : (int)cam_y;
+	cam_x = (float)((cam_x - (int)cam_x >= 0.5f) ? (int)cam_x + 1 : (int)cam_x);
+	cam_y = (float)((cam_y - (int)cam_y >= 0.5f) ? (int)cam_y + 1 : (int)cam_y);
 
 	D3DXVECTOR3 p(x - cam_x, y - cam_y, 0);
 	RECT r; 
@@ -346,8 +346,8 @@ bool CGame::CheckInCameraState(CGameObject* object)
 	CCamera::GetInstance()->GetPosition(cx, cy);
 
 	// Get screen width and height
-	float sw = GetScreenWidth();
-	float sh = GetScreenHeight();
+	float sw = (float)GetScreenWidth();
+	float sh = (float)GetScreenHeight();
 
 	return !(x > cx + sw || x + width < cx || y > cy + sh || y + height < cy);
 }
