+++ mario-directx/Sprites.cpp
@@ -25,8 +25,8 @@ void CSprite::Draw(float x, float y, int alpha)
 	CGame * game = CGame::GetInstance();
 	
 	// Fix rendering in order to avoid 'real value' render position
-	game->Draw((x - (int)x >= 0.5f) ? (int)x + 1 : (int)x,
-		(y - (int)y >= 0.5f) ? (int)y + 1 : (int)y, texture, left, top, right, bottom, alpha);
+	game->Draw((float)((x - (int)x >= 0.5f) ? (int)x + 1 : (int)x),
+		(float)((y - (int)y >= 0.5f) ? (int)y + 1 : (int)y), texture, left, top, right, bottom, alpha);
 }
 
 void CSprites::Add(int id, int left, int top, int right, int bottom, LPDIRECT3DTEXTURE9 tex)
